requirements.txt
tensorflow
streamlit
scikit-learn
matplotlib
numpy
Pillow
